CLEAN code

+++ src/services/dao/ServicesDao.js
@@ -3,7 +3,6 @@ const {Service} = require("../models/Service");
 const {Week} = require("../models/Week");
 const {CalendarDate} = require("../../utils/Date.utils");
 
-const currentService = 'actualService';
 const currentWeek = 'week';
 
 
@@ -15,7 +14,9 @@ const ServicesDao = {
         const weekToStore =
             {
                 services: week.services
-                    .map(service => ({...service, date:service.date.toISODateString()}))
+                    .map(service => (
+                        {...service, date:service.date.toISODateString()}
+                    ))
             }
 
         db.setItem(currentWeek,JSON.stringify(weekToStore))
@@ -23,19 +24,19 @@ const ServicesDao = {
 
     getWeek: () => {
         const store = db.getItem(currentWeek);
-        if (store === null){
-            return null;
-        }
+        if (store === null) return null;
 
         const servicesFromDatabase = JSON.parse(store).services;
 
         const services = servicesFromDatabase.map(element => new Service(element.code, new CalendarDate(element.date)))
 
         return new Week(services);
+
     },
 
     cleanWeek: () => {
-        db.removeItem(currentWeek);
+        if(db.getItem(currentWeek) !== null) db.removeItem(currentWeek);
+
     }
 }
